Modified type_conversion.py

+++ BasicsOfPython/type_conversion.py
@@ -1,9 +1,9 @@
 # Type Conversion
 
-a = int("2")
+a = float("2")
 b = 4.25
 
 sum = a + b # 2.0 + 4.25 => 6.25
 
 print(type(a))
-print(sum)
+print(sum)
